requirements.txt
Flask==3.0.3
flask-cors==5.0.0
gunicorn==22.0.0
authlib==1.3.1
google-cloud-bigquery==3.25.0
requests==2.32.3
